chore: move generate report markdown to separate package
