+++ packages/bundlemon-markdown-output/README.md
@@ -1,0 +1,5 @@
+# bundlemon-markdown-output
+
+[![npm](https://img.shields.io/npm/v/bundlemon-markdown-output)](http://www.npmjs.com/package/bundlemon-markdown-output)
+
+**Full documentation available [here](https://github.com/LironEr/bundlemon)**
