+++ packages/bundlemon-markdown-output/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "bundlemon-utils",
-  "version": "0.4.0",
+  "name": "bundlemon-markdown-output",
+  "version": "0.1.0",
   "description": "",
   "keywords": [],
   "author": "Liron Er",
@@ -9,13 +9,12 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/LironEr/bundlemon.git",
-    "directory": "packages/bundlemon-utils"
+    "directory": "packages/bundlemon-markdown-output"
   },
   "main": "./lib/cjs/index.js",
   "module": "./lib/esm/index.js",
   "types": "./lib/esm/index.d.ts",
   "scripts": {
-    "test": "jest",
     "build": "rimraf lib/ && tsc -p tsconfig.release.json && tsc -p tsconfig-cjs.json",
     "prepublishOnly": "yarn test && yarn lint && yarn build",
     "lint": "yarn eslint --config ../../.eslintrc.json --max-warnings=0 \"src/**/*.ts\"",
@@ -23,7 +22,8 @@
     "type-check": "tsc --noEmit"
   },
   "dependencies": {
-    "bytes": "^3.1.0"
+    "bytes": "^3.1.0",
+    "bundlemon-utils": "^0.4.0"
   },
   "devDependencies": {
     "@types/bytes": "^3.1.0",
