+++ packages/bundlemon-markdown-output/src/index.ts
@@ -1,0 +1,2 @@
+export { generateReportMarkdown } from './generator';
+export * from './types';
