+++ packages/bundlemon-markdown-output/src/types.ts
@@ -1,0 +1,3 @@
+export interface GenerateReportMarkdownOptions {
+  showLinkToReport?: boolean;
+}
