+++ packages/bundlemon-markdown-output/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "outDir": "lib/esm",
+    "rootDir": "src"
+  },
+  "include": ["src/**/*"]
+}
