+++ packages/bundlemon-markdown-output/tsconfig.release.json
@@ -1,0 +1,4 @@
+{
+  "extends": "./tsconfig.json",
+  "exclude": ["**/__tests__"]
+}
