+++ packages/bundlemon/src/index.ts
@@ -1,0 +1,1 @@
+export { default } from './main';
